agent/requirements.txt
livekit
livekit-api
livekit-agents
livekit-plugins-groq
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-deepgram
livekit-plugins-sarvam
livekit-plugins-elevenlabs
langgraph>=0.2.0
python-dotenv>=0.19.0
requests>=2.26.0
google
google-genai
